fix: restore valid socket server code;
