+++ packages/sdk-socket-server/api-config.js
@@ -1,18 +1,17 @@
 /* eslint-disable node/no-process-env */
-import crypto from 'crypto';
-import express from 'express';
-import bodyParser from 'body-parser';
-import cors from 'cors';
-import helmet from 'helmet';
-import { LRUCache } from 'lru-cache';
-import Analytics from 'analytics-node';
+const crypto = require('crypto');
+const express = require('express');
+const bodyParser = require('body-parser');
+const cors = require('cors');
+const helmet = require('helmet');
+const { LRUCache } = require('lru-cache');
+const Analytics = require('analytics-node');
 
-const isDevelopment: boolean = process.env.NODE_ENV === 'development';
-const isDevelopmentServer: boolean = process.env.ENVIRONMENT === 'development';
+const isDevelopment = process.env.NODE_ENV === 'development';
 
-const userIdHashCache = new LRUCache<string, string>({
+const userIdHashCache = new LRUCache({
   max: 5000,
-  ttl: 1000 * 60 * 60 * 24,
+  maxAge: 1000 * 60 * 60 * 24,
 });
 
 const app = express();
@@ -25,12 +24,12 @@ app.use(helmet());
 app.disable('x-powered-by');
 
 const analytics = new Analytics(
-  isDevelopment || isDevelopmentServer
-    ? process.env.SEGMENT_API_KEY_DEBUG || ''
-    : process.env.SEGMENT_API_KEY_PRODUCTION || '',
+  isDevelopment
+    ? process.env.SEGMENT_API_KEY_DEBUG
+    : process.env.SEGMENT_API_KEY_PRODUCTION,
   {
     flushInterval: isDevelopment ? 1000 : 10000,
-    errorHandler: (err: Error) => {
+    errorHandler: (err) => {
       console.error(`ERROR> Analytics-node flush failed: ${err}`);
     },
   },
@@ -52,22 +51,15 @@ app.post('/debug', (_req, res) => {
       return res.status(400).json({ error: 'wrong event name' });
     }
 
-    const id: string = body.id || 'socket.io-server';
-    let userIdHash: string | undefined = userIdHashCache.get(id);
+    const id = body.id || 'socket.io-server';
+    let userIdHash = userIdHashCache.get(id);
 
     if (!userIdHash) {
       userIdHash = crypto.createHash('sha1').update(id).digest('hex');
       userIdHashCache.set(id, userIdHash);
     }
 
-    const event: {
-      userId: string;
-      event: string;
-      properties: {
-        userId: string;
-        [key: string]: string | undefined; // This line adds an index signature
-      };
-    } = {
+    const event = {
       userId: userIdHash,
       event: body.event,
       properties: {
@@ -76,7 +68,7 @@ app.post('/debug', (_req, res) => {
     };
 
     // Define properties to be excluded
-    const propertiesToExclude: string[] = ['icon'];
+    const propertiesToExclude = ['icon'];
 
     for (const property in body) {
       if (
@@ -92,9 +84,9 @@ app.post('/debug', (_req, res) => {
       console.log('DEBUG> Event object:', event);
     }
 
-    analytics.track(event, function (err: Error) {
+    analytics.track(event, function (err, batch) {
       if (isDevelopment) {
-        console.log('DEBUG> Segment batch');
+        console.log('DEBUG> Segment batch', batch);
       }
 
       if (err) {
@@ -108,4 +100,4 @@ app.post('/debug', (_req, res) => {
   }
 });
 
-export { app, analytics };
+module.exports = { app, analytics };
