+++ packages/sdk-socket-server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@metamask/sdk-socket-server",
-  "version": "2.0.0",
+  "version": "1.0.0",
   "private": true,
   "description": "",
   "homepage": "https://github.com/MetaMask/metamask-sdk#readme",
@@ -12,22 +12,21 @@
     "url": "https://github.com/MetaMask/metamask-sdk.git"
   },
   "author": "",
-  "main": "dist/index.js",
+  "main": "index.js",
   "scripts": {
-    "build": "tsc",
+    "build": "echo 'N/A'",
     "build:post-tsc": "echo 'N/A'",
     "build:pre-tsc": "echo 'N/A'",
-    "clean": "rimraf dist",
-    "debug": "NODE_ENV=development ts-node-dev --respawn --transpile-only --inspect -- index.ts",
+    "clean": "echo 'N/A'",
+    "debug": "NODE_ENV=development nodemon index.js --trace-warnings",
     "lint": "yarn lint:eslint && yarn lint:misc --check",
     "lint:changelog": "yarn auto-changelog validate",
     "lint:eslint": "eslint . --cache --ext js,ts",
     "lint:fix": "yarn lint:eslint --fix && yarn lint:misc --write",
     "lint:misc": "prettier '**/*.json' '**/*.md' '!CHANGELOG.md' --ignore-path ../../.gitignore",
     "reset": "yarn clean && rimraf ./node_modules/",
-    "start": "yarn build && node dist/packages/sdk-socket-server/index.js",
-    "start:docker": "docker run -p 4000:4000 -d socket-test",
-    "test": "jest",
+    "start": "node index.js",
+    "test": "echo 'N/A'",
     "test:ci": "jest --coverage --passWithNoTests"
   },
   "dependencies": {
@@ -46,19 +45,6 @@
   "devDependencies": {
     "@lavamoat/allow-scripts": "^2.3.1",
     "@metamask/auto-changelog": "^2.3.0",
-    "@types/analytics-node": "^3.1.13",
-    "@types/body-parser": "^1.19.4",
-    "@types/cors": "^2.8.15",
-    "@types/dotenv": "^8.2.0",
-    "@types/express": "^4.17.20",
-    "@types/helmet": "^4.0.0",
-    "@types/jest": "^29.5.6",
-    "@types/node": "^20.8.7",
-    "@types/socket.io": "^3.0.2",
-    "@types/supertest": "^2.0.15",
-    "@types/uuid": "^8.3.1",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.8.0",
     "eslint": "^7.30.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.4",
@@ -68,13 +54,7 @@
     "eslint-plugin-prettier": "^3.4.0",
     "jest": "^29.6.4",
     "nodemon": "^2.0.20",
-    "prettier": "^2.8.8",
-    "socket.io-client": "^4.7.2",
-    "supertest": "^6.3.3",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^5.2.2"
+    "prettier": "^2.8.8"
   },
   "lavamoat": {
     "allowScripts": {
