--- packages/sdk-socket-server/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES6",
-    "module": "commonjs",
-    "outDir": "dist",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
-  },
-  "include": ["*.ts"],
-  "exclude": [
-    "node_modules",
-    "*.spec.ts",
-    "*.test.ts",
-    "*.spec.tsx",
-    "*.test.tsx"
-  ]
-}
